requirements.txt
Django==3.0.5
djangorestframework==3.11.0
djangorestframework_simplejwt==4.1.5
django-configurations==2.2
psycopg2-binary